Refactored Positions
Refactored Positions to have just the title of a qualification as the requirements, not an actual qualification object. Overloaded the == operator to compare Qualification objects with strings of their Name. Updated tests for comparisons to include expiration dates. Fleshed out qualification expiration. Added "CanExpire" flag to check if an expiration date is needed, or needs to be checked.